pyproject.toml
[project]
name = "bookmark-tagger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
